adp/dataflow/coderunner/coderunner/requirements.txt
tornado==6.5
RestrictedPython==8.0
pyparsing==3.1.1
openpyxl==3.0.10
jsonschema==4.20.0
xlrd==2.0.1
xlwt==1.3.0
python-docx==1.1.0
xlutils==2.0.0
requests==2.31.0
httpx==0.26.0
urllib3==2.6.0
python-dateutil==2.8.2
pydantic==2.5.3
annotated-types==0.6.0
pyahocorasick==2.1.0
pandas==2.0.3
PyMuPDF==1.24.2
pyzbar==0.1.9
pdfplumber==0.11.8
docxtpl==0.16.8
opencv-python==4.9.0.80
PyMySQL==1.1.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
numpy==1.24.4
